fixups and refactor
