+++ docs/package.json
@@ -27,17 +27,15 @@
   },
   "dependencies": {
     "@docusaurus/core": "^2.0.0-beta.14",
-    "@docusaurus/plugin-google-analytics": "^2.0.0-beta.14",
+    "@docusaurus/plugin-ideal-image": "^2.0.0-beta.14",
     "@docusaurus/preset-classic": "^2.0.0-beta.14",
     "@mdx-js/react": "^1.6.21",
     "clsx": "^1.1.1",
+    "docusaurus-plugin-sass": "^0.2.1",
     "prism-react-renderer": "^1.2.1",
     "react": "^17.0.1",
-    "react-dom": "^17.0.1"
-  },
-  "resolutions": {
-    "nth-check": "^2.0.1",
-    "trim": "0.0.3"
+    "react-dom": "^17.0.1",
+    "sass": "^1.49.0"
   },
   "devDependencies": {
     "prettier": "^2.5.1"
