+++ docs/src/pages/index.js
@@ -4,44 +4,46 @@ import Layout from '@theme/Layout'
 import Link from '@docusaurus/Link'
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext'
 import useBaseUrl from '@docusaurus/useBaseUrl'
-import styles from './styles.module.css'
-import InfoModal from './components/InfoModal'
+import styles from './styles.module.scss'
+import InfoModal from '../components/InfoModal'
 
 function Home() {
   const context = useDocusaurusContext()
   const { siteConfig = {} } = context
   return (
-    <Layout
-      title={`Approzium`}
-      description="Enhance the observability and security of your database applications"
-    >
-      <header className={clsx('hero hero--primary', styles.heroBanner)}>
-        <div className={clsx('container', styles.container)}>
-          <div>
-            <img className={clsx('icon', styles.icon)} src={'img/apzm-icon.svg'} />
-            <h1 className={clsx('hero__title', styles.heroTitle)}>{siteConfig.title}</h1>
-            <p className={clsx('hero__subtitle', styles.heroSubtitle)}>
-              Enhance the <strong>observability</strong> and <strong>security</strong>
-              <br /> of your database applications.
-            </p>
+    <div className={styles.homepageWrapper}>
+      <Layout
+        title="Approzium"
+        description="Enhance the observability and security of your database applications"
+      >
+        <header className={clsx('hero hero--primary', styles.heroBanner)}>
+          <div className={clsx('container', styles.container)}>
+            <div>
+              <img className={clsx('icon', styles.icon)} src={'img/apzm-icon.svg'} />
+              <h1 className={clsx('hero__title', styles.heroTitle)}>{siteConfig.title}</h1>
+              <p className={clsx('hero__subtitle', styles.heroSubtitle)}>
+                Enhance the <strong>observability</strong> and <strong>security</strong>
+                <br /> of your database applications.
+              </p>
+            </div>
+            <div className={styles.buttons}>
+              <Link
+                className={clsx(
+                  'button button--outline button--secondary button--lg',
+                  styles.getStarted
+                )}
+                to={useBaseUrl('docs/')}
+              >
+                Get Started
+              </Link>
+            </div>
           </div>
-          <div className={styles.buttons}>
-            <Link
-              className={clsx(
-                'button button--outline button--secondary button--lg',
-                styles.getStarted
-              )}
-              to={useBaseUrl('docs/')}
-            >
-              Get Started
-            </Link>
-          </div>
-        </div>
-      </header>
-      <main>
-        <InfoModal />
-      </main>
-    </Layout>
+        </header>
+        <main>
+          <InfoModal />
+        </main>
+      </Layout>
+    </div>
   )
 }
