+++ docs/src/pages/styles.module.scss
@@ -26,6 +26,16 @@
   background-color: var(--ifm-background-color);
 }
 
+.homepageWrapper {
+  nav {
+    border-bottom: none;
+  }
+
+  footer {
+    border-top: none;
+  }
+}
+
 @media screen and (max-width: 966px) {
   .heroBanner {
     padding: 2rem;
