--- docs/src/theme/CodeBlock/styles.module.css
@@ -1,77 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-.codeBlockContent {
-  position: relative;
-  /*rtl:ignore*/
-  direction: ltr;
-}
-
-.codeBlockTitle {
-  border-top-left-radius: var(--ifm-global-radius);
-  border-top-right-radius: var(--ifm-global-radius);
-  border-bottom: 1px solid var(--ifm-color-emphasis-200);
-  font-family: var(--ifm-font-family-monospace);
-  font-weight: bold;
-  padding: 0.75rem var(--ifm-pre-padding);
-  width: 100%;
-}
-
-.codeBlock {
-  overflow: auto;
-  border-radius: var(--ifm-global-radius);
-}
-
-.codeBlockContainer {
-  margin-bottom: var(--ifm-leading);
-  border-radius: var(--ifm-global-radius);
-  box-shadow: var(--ifm-global-shadow-lw);
-}
-
-.codeBlockTitle + .codeBlockContent .codeBlock {
-  border-top-left-radius: 0;
-  border-top-right-radius: 0;
-}
-
-.copyButton {
-  background: rgba(0, 0, 0, 0.3);
-  border: none;
-  border-radius: var(--ifm-global-radius);
-  color: var(--ifm-color-white);
-  cursor: pointer;
-  opacity: 0;
-  outline: none;
-  padding: 0.4rem 0.5rem;
-  position: absolute;
-  right: calc(var(--ifm-pre-padding) / 2);
-  top: calc(var(--ifm-pre-padding) / 2);
-  visibility: hidden;
-  transition: opacity 200ms ease-in-out, visibility 200ms ease-in-out, bottom 200ms ease-in-out;
-}
-
-.codeBlockTitle:hover + .codeBlockContent .copyButton,
-.codeBlockContent:hover > .copyButton,
-.copyButton:focus {
-  visibility: visible;
-  opacity: 1;
-}
-
-.codeBlockLines {
-  font-family: var(--ifm-font-family-monospace);
-  font-size: inherit;
-  line-height: var(--ifm-pre-line-height);
-  white-space: pre;
-  float: left;
-  min-width: 100%;
-  padding: var(--ifm-pre-padding);
-}
-
-@media print {
-  .codeBlockLines {
-    white-space: pre-wrap;
-  }
-}
