--- docs/src/theme/Footer/styles.module.css
@@ -1,19 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-.footerLogoLink {
-  opacity: 0.5;
-  transition: opacity var(--ifm-transition-fast) var(--ifm-transition-timing-default);
-}
-
-.footerLogoLink:hover {
-  opacity: 1;
-}
-
-.footerBorder {
-  border-top: 1px solid var(--ifm-toc-border-color);
-}
