--- docs/src/theme/Navbar/styles.module.css
@@ -1,31 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-/*
-Hide toggle in small viewports
- */
-@media (max-width: 996px) {
-  .toggle {
-    display: none;
-  }
-}
-
-.navbarHideable {
-  transition: transform var(--ifm-transition-fast) ease;
-}
-
-.navbarHidden {
-  transform: translate3d(0, calc(-100% - 2px), 0);
-}
-
-.navbarSidebarToggle {
-  margin-right: 1rem;
-}
-
-.hideBorder {
-  border-bottom: none;
-}
